Separate 'NewCommandView' from 'EditCommandView'
