+++ Source/SelfService/Web/integrations/connection/connectionDetails/commands/command/newCommand/NewCommandDetailSection.tsx
@@ -5,9 +5,9 @@ import React from 'react';
 
 import { ContentWithSubtitle, InlineWrapper, Input } from '@dolittle/design-system';
 
-import { alphaNumericCharsRegex } from '../../../../../utils/helpers/regex';
+import { alphaNumericCharsRegex } from '../../../../../../utils/helpers/regex';
 
-export const CommandDetailSection = () =>
+export const NewCommandDetailSection = () =>
     <ContentWithSubtitle title='Command details' infoTooltipLabel='Provide a name, namespace and description for your command. Command name is required.'>
         <InlineWrapper sx={{ alignItems: 'flex-start', gap: 4 }}>
             <Input id='commandName' label='Command name' required pattern={{ value: alphaNumericCharsRegex, message: 'Can only contain characters or numbers.' }} />
