+++ Source/SelfService/Web/integrations/connection/connectionDetails/commands/command/newCommand/commandSection/TransactionsDataGridColumns.tsx
